backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
langchain
langgraph
openai
python-dotenv
supabase
